Move all integration tests into /route
