+++ src/routes/upload_test.go
@@ -1,4 +1,4 @@
-package integrationtests
+package routes
 
 import (
 	"bytes"
